added style and more functionality

+++ product.php
@@ -13,19 +13,20 @@ if (!isset($_SESSION["admin"])) {
 $product_title = '';
 $product_description = '';
 $product_price = '';
+$params = ['title', 'description', 'price', 'image'];
 
 switch ($_GET['action']){
     case 'create':
         if(isset($_POST['submit'])){
             // Check if image file is a actual image or fake image
-            $image_checked = imageValidator();
+            $image_checked = imageValidator($image_validator_errors);
 
             if($image_checked['upload_ok'])
             {
                 $sql = "INSERT INTO `products` (`title`, `description`, `price`, `image`) VALUES (?,?,?,?)";
                 $stmt = $connect_db->prepare($sql);
                 $stmt->bind_param("ssds", $_POST['title'], $_POST['description'], $_POST['price'], $image_checked["image_name"]);
-                echo $stmt->execute() ?  "The product was inserted in db" : "There was an error on trying to insert the new product in db";
+                echo $stmt->execute() ?  $product_page['product_created'] : $product_page['product_error'];
             }
         }
         break;
@@ -43,33 +44,36 @@ switch ($_GET['action']){
 
             if(isset($_POST["submit"]))
             {
-                $image_checked = imageValidator();
+                $image_checked = imageValidator($image_validator_errors);
                 $id = mysqli_real_escape_string($connect_db,$id);
                 $sql = "UPDATE `products`  SET `title`=?, `description`=?, `price`=?, `image`=? WHERE `id`='" . $id . "'";
                 $stmt = $connect_db->prepare($sql);
          
-                $product_image = $image_checked['upload_ok'] ? 
-                $image_checked["image_name"] : $product_image;
+                if($image_checked['upload_ok'])
+                {
+                    unlink($product_image);
+                    $product_image = $image_checked["image_name"];
+                }
 
-                $stmt->bind_param("ssds", 
-                    $_POST['title'], 
-                    $_POST['description'], 
-                    $_POST['price'], 
+                $stmt->bind_param("ssds",
+                    $_POST['title'],
+                    $_POST['description'],
+                    $_POST['price'],
                     $product_image
                 );
-                echo $stmt->execute() ?  $product_page['good_message'] : $product_page['wrong_message'];
+                echo $stmt->execute() ?  $product_page['update_product'] : $product_page['update_product_error'];
             }
                
         }
        
         break;
     case 'delete':
-        if(isset($_GET['id'])){
+        if(isset($_GET['id']) && isset($_GET['image'])){
             $id = $_GET['id'];
             $id = mysqli_real_escape_string($connect_db,$id);
             $sql = "DELETE FROM `products` WHERE `id`='" . $id . "'";
             $connect_db->query($sql);
-           
+            unlink($_GET['image']);
             header("Location: http://localhost/products.php");
         }
 
@@ -82,14 +86,14 @@ switch ($_GET['action']){
 <div class="container">
 
 <form action="<?php echo $_SERVER['REQUEST_URI'] ?>" method="post" class="form-group" enctype="multipart/form-data">
-    <input type="text" placeholder="<?php echo $product_page['form_title'] ?>" value="<?php echo $product_title ?>" name="title" class="form-control" required>
-    <input type="text" placeholder="<?php echo $product_page['form_description'] ?>" value="<?php echo $product_description ?>" name="description" class="form-control" required>
+    <input type="text" placeholder="<?php echo $product_page['form_title'] ?>" value="<?php echo $product_title ?>" name="title" class="form-control" style='margin:10px 0' required>
+    <input type="text" placeholder="<?php echo $product_page['form_description'] ?>" value="<?php echo $product_description ?>" name="description" class="form-control" style='margin:10px 0' required>
 
-    <input type="number" step="any" placeholder="<?php echo $product_page['form_price'] ?>" value="<?php echo $product_price ?>"name="price" class="form-control" required>
+    <input type="number" step="any" placeholder="<?php echo $product_page['form_price'] ?>" value="<?php echo $product_price ?>"name="price" class="form-control" style='margin:10px 0' required>
     
-    <input type="file" name="fileToUpload" class="form-control-file" id="formFile">
+    <input type="file" name="fileToUpload" class="form-control-file" id="formFile" <?php if($_GET['action'] === "create") echo "required" ?>  >
 
-    <a href="products.php"  name="submit"><?php echo $product_page['products'] ?></a>
-    <button type="submit" value="click"  name="submit"><?php echo $product_page['save'] ?></button>
+    <a href="products.php"  name="submit" class='btn btn-dark' style='margin: 10px'><?php echo $product_page['products'] ?></a>
+    <button type="submit" value="click"  name="submit" class='btn btn-primary' style='margin: 10px'><?php echo $product_page['save'] ?></button>
 </form>
 </div>
